mcp-server/requirements.txt
fastmcp>=0.1.0
psycopg2-binary>=2.9.9
pydantic>=2.8.0
